Working changes. Implementing A* search

+++ src/Entity.cpp
@@ -1,12 +1,12 @@
 #include "Entity.h"
 
-Entity::Entity(int id, float x, float y)
+Entity::Entity(int id, double x, double y)
 {
     _id = id;
     this->setLocation(x, y);
 }
 
-void Entity::setLocation(float x, float y)
+void Entity::setLocation(double x, double y)
 {
     _x = x;
     _y = y;
